power: reset: msm-poweroff: Restructure restart reasons for SoMC
This change restructures the restart reasons handling for SoMC
devices through the introduction of the definition of a generic
TARGET_SOMC_BOOTLOADER, which is valid for both S1BOOT and XBOOT
powered devices.
These bootloaders do need a specific restart reason to reboot to
at least recovery environment (boot selection to recovery partition)
which is different from the qcom's standard definition.